+++ wiggle_shell/core/select_dbms/tests/test_select_index_file.py
@@ -4,7 +4,7 @@ from typing import Generator
 import pytest
 
 from testing import TEST_DBMS_FOLDER_PATH
-from wiggle_shell.core.select_dbms.select_relationship_index import (
+from wiggle_shell.core.select_dbms.select_index_file import (
     create_new_relationship_index,
     get_existing_relationship_index_file_path,
 )
